Rename component/base_classes
