+++ src/components/aocs/initialize_gyro_sensor.cpp
@@ -6,7 +6,7 @@
 
 #include <interface/initialize/initialize_file_access.hpp>
 
-#include "../base_classes/initialize_sensor_base.hpp"
+#include "../base/initialize_sensor_base.hpp"
 
 Gyro InitGyro(ClockGenerator* clock_gen, int sensor_id, const std::string fname, double compo_step_time, const Dynamics* dynamics) {
   IniAccess gyro_conf(fname);
