+++ src/components/aocs/initialize_magnetometer.cpp
@@ -4,7 +4,7 @@
  */
 #include "initialize_magnetometer.hpp"
 
-#include "../base_classes/initialize_sensor_base.hpp"
+#include "../base/initialize_sensor_base.hpp"
 #include "interface/initialize/initialize_file_access.hpp"
 
 MagSensor InitMagSensor(ClockGenerator* clock_gen, int sensor_id, const std::string fname, double compo_step_time, const MagEnvironment* magnet) {
