+++ src/components/aocs/magnetometer.hpp
@@ -10,8 +10,8 @@
 #include <interface/log_output/loggable.hpp>
 #include <library/math/quaternion.hpp>
 
-#include "../base_classes/component_base.hpp"
-#include "../base_classes/sensor_base.hpp"
+#include "../base/component_base.hpp"
+#include "../base/sensor_base.hpp"
 
 const size_t kMagDim = 3;  //!< Dimension of magnetometer
