+++ src/components/aocs/star_sensor.hpp
@@ -15,7 +15,7 @@
 #include <library/math/vector.hpp>
 #include <vector>
 
-#include "../base_classes/component_base.hpp"
+#include "../base/component_base.hpp"
 #include "dynamics/dynamics.hpp"
 
 /*
