+++ src/components/base/component_base.cpp
