+++ src/components/examples/example_change_structure.hpp
@@ -9,7 +9,7 @@
 #include <interface/log_output/loggable.hpp>
 #include <simulation/spacecraft/structure/structure.hpp>
 
-#include "../base_classes/component_base.hpp"
+#include "../base/component_base.hpp"
 
 /**
  * @class ExampleChangeStructure
