+++ src/components/examples/example_i2c_target_for_hils.hpp
@@ -8,8 +8,8 @@
 
 #include <vector>
 
-#include "../base_classes/component_base.hpp"
-#include "../base_classes/obc_i2c_target_communication_base.hpp"
+#include "../base/component_base.hpp"
+#include "../base/obc_i2c_target_communication_base.hpp"
 
 /**
  * @class ExampleI2cTargetForHils
