+++ src/components/examples/example_serial_communication_for_hils.hpp
@@ -8,8 +8,8 @@
 
 #include <vector>
 
-#include "../base_classes/component_base.hpp"
-#include "../base_classes/obc_communication_base.hpp"
+#include "../base/component_base.hpp"
+#include "../base/obc_communication_base.hpp"
 
 /**
  * @class ExampleSerialCommunicationForHils
