+++ src/components/examples/example_serial_communication_with_obc.hpp
@@ -8,9 +8,9 @@
 
 #include <vector>
 
-#include "../base_classes/component_base.hpp"
-#include "../base_classes/interface_gpio_component.hpp"
-#include "../base_classes/obc_communication_base.hpp"
+#include "../base/component_base.hpp"
+#include "../base/interface_gpio_component.hpp"
+#include "../base/obc_communication_base.hpp"
 
 /**
  * @class ExampleSerialCommunicationWithObc
