+++ src/components/mission/telescope.hpp
@@ -6,7 +6,7 @@
 #ifndef S2E_COMPONENTS_MISSION_TELESCOPE_HPP_P_
 #define S2E_COMPONENTS_MISSION_TELESCOPE_HPP_P_
 
-#include <base_classes/component_base.hpp>
+#include <base/component_base.hpp>
 #include <dynamics/attitude/attitude.hpp>
 #include <environment/global/hipparcos_catalogue.hpp>
 #include <environment/local/local_celestial_information.hpp>
